requirements.txt
flask
flask-cors
numpy
pandas
torch
torchvision
scikit-learn
xgboost
pillow
gunicorn